+++ src/App.js
@@ -2,14 +2,15 @@ import "./App.css";
 import Navbar from "./components/toolbar/Navbar.jsx";
 import Footer from "./components/toolbar/Footer.jsx";
 import { Routes, Route, useLocation } from "react-router-dom";
-import LoginPage from "./pages/LoginPage.js";
+import LoginPage from "./pages/LoginPage.jsx";
 import Main from "./pages/Main.js";
 import Analytics from "./pages/Analytics.js";
 import Calendar from "./pages/Calendar.js";
 import Daily from "./pages/Daily.js";
 import Recommend from "./pages/Recommend.js";
-import Signin from "./pages/SigninPage.js";
-import Mypage from "./pages/Mypage.js";
+import Signin from "./pages/SigninPage.jsx";
+import Nickname from "./components/signin/Nickname.jsx";
+import Mypage from "./pages/Mypage.jsx";
 
 function App() {
   const location = useLocation();
@@ -18,12 +19,13 @@ function App() {
   return (
     <div>
       <Navbar />
-      <Routes> 
+      <Routes>
         <Route path="/" element={<Main />} />
         <Route path="/analytics" element={<Analytics />} />
         <Route path="/calander" element={<Calendar />} />
         <Route path="/login" element={<LoginPage />} />
         <Route path="/login/signin" element={<Signin />} />
+        <Route path="/login/signin/nickname" element={<Nickname />} />
         <Route path="/daily" element={<Daily />} />
         <Route path="/recommend" element={<Recommend />} />
         <Route path="/mypage" element={<Mypage />} />
