+++ src/components/login/LoginButton.jsx
@@ -2,17 +2,14 @@ import React from "react";
 import styled from "styled-components";
 import theme from "../../styles/theme";
 
-const LoginButton = ({children}) => {
-  return (
-    <Login>{children}</Login>
-  );
+const LoginButton = ({ children }) => {
+  return <Login>{children}</Login>;
 };
 
 const Login = styled.button`
-  width: 46rem;
+  width: 46.7rem;
   height: 6.5rem;
   background-color: ${theme.colors["main-purple"]};
-  margin-top: 7rem;
   color: white;
   font-size: 2.5rem;
   font-weight: bold;
